Move converter factory to sub-package
